Fix rebase error
The commit "Add Support for Split in CascadingCommandStreamGenerator"
accidentally undid the fix in commit "Fix uint8 overflow".

This caused an error when running larger layers with >256 stripes as the
number of stripes overflow the underlying datatype when calculating the
dependency ratios.

+++ driver/support_library/src/cascading/CascadingCommandStreamGeneratorUtils.hpp
@@ -659,12 +659,12 @@ inline uint16_t CalculateGCD(uint16_t a, uint16_t b)
 {
     if (a == 0)
     {
-        return ethosn::utils::NumericCast<uint8_t>(b);
+        return ethosn::utils::NumericCast<uint16_t>(b);
     }
     return CalculateGCD(b % a, a);
 }
 
-inline uint16_t FindGreatestCommonDenominator(uint16_t a, uint16_t b, uint8_t c)
+inline uint16_t FindGreatestCommonDenominator(uint16_t a, uint16_t b, uint16_t c)
 {
     uint16_t gcdAB = CalculateGCD(a, b);
     if (c == 0)
